# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/teachme-api/public/api",
    "paths": {
        "/anuncios": {
            "get": {
                "tags": [
                    "Anuncio"
                ],
                "summary": "Get a listing of the Anuncios.",
                "description": "Get all Anuncios",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Anuncio"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Anuncio"
                ],
                "summary": "Store a newly created Anuncio in storage",
                "description": "Store Anuncio",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Anuncio that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Anuncio"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Anuncio"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/anuncios/{id}": {
            "get": {
                "tags": [
                    "Anuncio"
                ],
                "summary": "Display the specified Anuncio",
                "description": "Get Anuncio",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Anuncio",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Anuncio"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Anuncio"
                ],
                "summary": "Update the specified Anuncio in storage",
                "description": "Update Anuncio",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Anuncio",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Anuncio that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Anuncio"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Anuncio"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Anuncio"
                ],
                "summary": "Remove the specified Anuncio from storage",
                "description": "Delete Anuncio",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Anuncio",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Anuncio": {
            "required": [
                "QtdAlunos",
                "CdDisciplina",
                "CdUsuario_Professor",
                "Valor"
            ],
            "properties": {
                "CdAnuncio": {
                    "description": "CdAnuncio",
                    "type": "integer",
                    "format": "int32"
                },
                "QtdAlunos": {
                    "description": "QtdAlunos",
                    "type": "integer",
                    "format": "int32"
                },
                "Descricao": {
                    "description": "Descricao",
                    "type": "string"
                },
                "CdDisciplina": {
                    "description": "CdDisciplina",
                    "type": "integer",
                    "format": "int32"
                },
                "CdUsuario_Professor": {
                    "description": "CdUsuario_Professor",
                    "type": "integer",
                    "format": "int32"
                },
                "Valor": {
                    "description": "Valor",
                    "type": "number",
                    "format": "number"
                }
            }
        },
...
    }
}